Stamp the guest-fn table via register_guest_kernel!
New guest_fns module carries the personality-generic bodies of all six
production RPCs (nub_invoke_cached/worker, put_object, evict_jit_all,
get_boot_info, heap_stats) plus the register_guest_kernel! macro that
emits the concrete #[guest_function] wrappers for one personality.
guest_prod reduces to a single invocation for Javm — the wrapper fn
idents are unchanged, so the linkme registration statics and fn_ids
are identical to before, and a second invocation in one crate is a
compile error (the structural one-personality guarantee). The interim
state_cache::evict_jit_all free fn folds into JavmStore::evict_jit.
