+++ rust/nub-arch-x86/src/lib.rs
@@ -33,6 +33,8 @@ pub mod call_loop;
 #[cfg(target_os = "none")]
 pub mod execution_lane;
 #[cfg(target_os = "none")]
+pub mod guest_fns;
+#[cfg(target_os = "none")]
 pub mod jit_cache;
 #[cfg(target_os = "none")]
 pub mod jit_run;
